separating card and grid functionality
