+++ src/app/components/pages/view-fish-page/view-fish-page.component.html
@@ -5,7 +5,7 @@
       </div>
       <div class="view-fish-page-content">
         <div class="display-grid-content">
-          <app-displaycard class="test"></app-displaycard>
+          <app-displaygrid class="test"></app-displaygrid>
         </div>
         <div class="pagination-content">
           <div class="prev-button">
