+++ src/app/components/ui/displaycard/displaycard.component.html
@@ -1,22 +1,9 @@
-<div class="grid-layout">
-
-  <div id="register-fish-popup"  *ngIf="show">
-    <app-register-fish
-    [fish]="selectedFish" id = "regfish"></app-register-fish>
+<div class="display-card-content">
+  <div class="overlap-group" style="background-image: url({{fish.image}});">
+    <img
+      class="name-border"
+      src="https://anima-uploads.s3.amazonaws.com/projects/6074ae6f11c6ad9a56067743/releases/60776f839e7a3de79356df1e/img/name-border@2x.png"
+    />
+    <h1 class="x-display-name">{{fish.Species}}</h1>
   </div>
-
-    <div class="display-card-content" *ngFor="let fish of fishPage" (click)="toggle(fish)">
-        <div class="overlap-group" style="background-image: url({{fish.image}});">
-          <img
-            class="name-border"
-            src="https://anima-uploads.s3.amazonaws.com/projects/6074ae6f11c6ad9a56067743/releases/60776f839e7a3de79356df1e/img/name-border@2x.png"
-          />
-          <h1 class="x-display-name">{{fish.Species}}</h1>
-        </div>
-      </div>
-      <mat-paginator [length] = "this.fishes.length"
-      [pageSize] = "15"
-      [pageIndex] = "0"
-      (page) = "onPageChange($event)"
-      ></mat-paginator>
 </div>
